Consolidate functions.
